+++ controllers/ConfigEditController.php
@@ -1,7 +1,0 @@
-<?php
-
-class ConfigEditController {
-    public function show() {
-
-    }
-}