+++ controllers/ConnexionController.php
@@ -1,0 +1,7 @@
+<?php
+
+class ConnexionController {
+    public function show() {
+        require_once 'website/luma/connexion.php';
+    }
+}