+++ controllers/HomeController.php
@@ -3,6 +3,7 @@
 
 class HomeController {
     public function index() {
+        $_GET['offline'] = false;
         require_once './website/luma/home.php';
     }
 }