+++ controllers/InitController.php
@@ -3,7 +3,6 @@
 
 class InitController {
     public function show() {
-        $titre = "Initialisation du site";
         require_once 'website/admin/init.php';
     }
 }